# repo/api/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CineRide Media Operations Platform API",
    "version": "1.0.0",
    "description": "REST API for the CineRide offline-first media operations platform. Manages movies, editorial content, ride requests, environmental sensors, and a local funds ledger."
  },
  "servers": [
    { "url": "/api", "description": "Local API (via Nginx reverse proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/movies": {
      "get": { "summary": "List movies", "tags": ["Movies"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "category", "in": "query", "schema": { "type": "string" } }, { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["popularity", "rating", "newest", "title"] } }], "responses": { "200": { "description": "Movie list" } } },
      "post": { "summary": "Create movie (staff only)", "tags": ["Movies"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Movie created" } } }
    },
...
    "/content": {
      "get": { "summary": "List content items", "tags": ["Content"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Content list" } } },
      "post": { "summary": "Create content (staff only)", "tags": ["Content"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Content created" } } }
    },
...
    "/rides": {
      "get": { "summary": "List ride requests", "tags": ["Rides"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Ride list" } } },
      "post": { "summary": "Create ride request", "tags": ["Rides"], "security": [{ "BearerAuth": [] }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "pickup_text": { "type": "string" }, "dropoff_text": { "type": "string" }, "rider_count": { "type": "integer", "minimum": 1, "maximum": 6 }, "time_window_start": { "type": "string", "format": "date-time" }, "time_window_end": { "type": "string", "format": "date-time" }, "vehicle_type": { "type": "string" } } } } } }, "responses": { "201": { "description": "Ride created" } } }
    },
...
    "/dispatch/queue": {
      "get": { "summary": "Get dispatcher queue (pending rides)", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Pending rides" } } }
    },
    "/dispatch/rides/{id}/accept": {
      "post": { "summary": "Accept/match a ride request", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Ride accepted" } } }
    },
    "/dispatch/disputes/{id}/resolve": {
      "post": { "summary": "Resolve a dispute", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Dispute resolved" } } }
    },
    "/disputes": {
      "post": { "summary": "Initiate a dispute (any role)", "tags": ["Disputes"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Dispute created" } } }
    },
...
    "/config": {
      "get": { "summary": "List config entries (admin only)", "tags": ["Config"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Config list" } } },
      "post": { "summary": "Create/update config entry", "tags": ["Config"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Config saved" } } }
    },
    "/extensions/clients": {
      "post": { "summary": "Create extension API client (admin only)", "tags": ["Extensions"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Client created with API key (shown once)" } } },
      "get": { "summary": "List extension clients", "tags": ["Extensions"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Client list" } } }
    },
    "/extensions/movies": {
      "get": { "summary": "Read movies via extension API", "tags": ["Extensions"], "security": [{ "ApiKeyAuth": [] }], "responses": { "200": { "description": "Movies" } } }
    },
...
    "/dispatch/rides/{id}/transition": {
      "post": { "summary": "Transition ride status (dispatcher)", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Ride transitioned" } } }
    },
    "/dispatch/rides/{id}/approve-cancel": {
      "post": { "summary": "Approve ride cancellation (dispatcher)", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Cancellation approved" } } }
    },
    "/dispatch/disputes": {
      "get": { "summary": "List disputes (dispatcher)", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "responses": { "200": { "description": "Dispute list" } } }
    },
    "/dispatch/disputes/{id}": {
      "get": { "summary": "Get dispute detail", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Dispute detail" } } }
    },
    "/dispatch/disputes/{id}/assign": {
      "post": { "summary": "Assign dispute to self", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Dispute assigned" } } }
    },
    "/dispatch/carpool/candidates/{id}": {
      "get": { "summary": "Get carpool candidates for a ride", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Candidate list" } } }
    },
    "/dispatch/carpool/group": {
      "post": { "summary": "Group rides into carpool", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "responses": { "201": { "description": "Carpool group created" } } }
    },
    "/dispatch/carpool/group/{groupId}": {
      "get": { "summary": "View carpool group", "tags": ["Dispatch"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "groupId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Carpool group rides" } } }
    },
...
    "/config/{key}": {
      "get": { "summary": "Get config value (admin)", "tags": ["Config"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Config value" } } },
      "put": { "summary": "Update config value (admin)", "tags": ["Config"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Config updated" } } },
      "delete": { "summary": "Delete config key (admin)", "tags": ["Config"], "security": [{ "BearerAuth": [] }], "parameters": [{ "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Config deleted" } } }
    },
    "/extensions/content": {
      "get": { "summary": "Read content via extension API", "tags": ["Extensions"], "security": [{ "ApiKeyAuth": [] }], "responses": { "200": { "description": "Content" } } }
    },
    "/extensions/rides": {
      "get": { "summary": "Read rides via extension API (minimized fields, filterable by status/date)", "tags": ["Extensions"], "security": [{ "ApiKeyAuth": [] }], "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending_match", "accepted", "in_progress", "completed", "canceled"] } }, { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } }, { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Rides (minimized safe field set)" }, "422": { "description": "Invalid status filter" } } }
    }
  }
}